% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{authorize}
\alias{authorize}
\title{Authorize R package to access Google Analytics API}
\usage{
authorize(client.id = getOption("gamanager.client.id"),
  client.secret = getOption("gamanager.client.secret"))
}
\arguments{
\item{client.id}{OAuth client ID. This is obtained from Google API Credentials}

\item{client.secret}{OAuth client secret. This is obtained from Google API Credentials}
}
\description{
This is a function to authorize the R package to access the Google Analytics API. If no
client.id and client.secret is provided, the package would provide predefined values.
}

