% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goals.R
\name{create_goal_visitNumPages}
\alias{create_goal_visitNumPages}
\title{Create a visit number of pages goals}
\usage{
create_goal_visitNumPages(accountId, webPropertyId, profileId, goalId, goalName,
  goalValue = NULL, active = TRUE, comparisonType, comparisonValue)
}
\arguments{
\item{accountId}{Account ID to retrieve profile for.}

\item{webPropertyId}{ID to retrieve the profile for.}

\item{profileId}{ID to retrive the profile for.}

\item{goalId}{The goal ID for the current specified goal.}

\item{goalName}{Goal name.}

\item{goalValue}{Goal value. Defaults to NULL}

\item{active}{Determines whether this goal is active. Defaults to TRUE}

\item{comparisonType}{Type of comparison. Possible values are LESS_THAN, GREATER_THAN, or EQUAL.}

\item{comparisonValue}{Value used for this comparison.}
}
\description{
Create a visit number of pages goals
}

