% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{create_profile}
\alias{create_profile}
\title{Create a new property in Google Analytics}
\usage{
create_profile(accountId, webPropertyId, name, botFilteringEnabled = NULL,
  currency = NULL, defaultPage = NULL, eCommerceTracking = NULL,
  enhancedECommerceTracking = NULL, excludeQueryParameters = NULL,
  type = "WEB", websiteUrl = NULL)
}
\arguments{
\item{accountId}{Account ID to link to new web property. Does not accept ~all as paramter}

\item{name}{Name of web property}

\item{websiteUrl}{Optional field. Website url for this web property.}

\item{industryVertical}{Optional field. The industry vertical/category selected for this web property.
If this field is set, the correct values are: UNSPECIFIED, ARTS_AND_ENTERTAINMENT}
}
\description{
Create a new property in Google Analytics
}

