% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamanager.R
\docType{package}
\name{gamanager}
\alias{gamanager}
\alias{gamanager-package}
\title{gamanager: Managing Google Analytics Management Settings via R}
\description{
Quick way to manage Google Analytics Managements settings with the gamanager package
}

