% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_endpoint}
\alias{get_endpoint}
\title{Generate endpoint for Google Analytics Management API
Temporarily available}
\usage{
get_endpoint(typeOfEndpoint = "gamanager.accounts.list", accountId = NULL,
  webPropertyId = NULL, profileId = NULL, goalId = NULL,
  customDimensionId = NULL, customMetricId = NULL)
}
\description{
Generate endpoint for Google Analytics Management API
Temporarily available
}

