% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goals.R
\name{get_goal}
\alias{get_goal}
\title{Get the details of a goal from Google Analytics}
\usage{
get_goal(accountId, webPropertyId, profileId, goalId)
}
\arguments{
\item{accountId}{Account ID to retrieve profile for.}

\item{webPropertyId}{ID to retrieve the profile for.}

\item{profileId}{ID to retrive the profile for.}

\item{goalId}{Goal ID to retrive the goal for.}
}
\description{
Get the details of a goal from Google Analytics
}

