% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{get_profile}
\alias{get_profile}
\title{Get the details of a web property from Google Analytics}
\usage{
get_profile(accountId, webPropertyId, profileId)
}
\arguments{
\item{accountId}{Account ID to retrieve profile for.}

\item{webPropertyId}{ID to retrieve the profile for.}

\item{profileId}{ID to retrive the profile for}
}
\description{
Get the details of a web property from Google Analytics
}

