% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_properties.R
\name{get_property}
\alias{get_property}
\title{Get the details of a web property from Google Analytics}
\usage{
get_property(accountId, webPropertyId)
}
\arguments{
\item{accountId}{Account ID to retrieve web properties for.}

\item{webPropertyId}{ID to retrieve the web property for.}
}
\description{
Get the details of a web property from Google Analytics
}

