% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{list_accounts}
\alias{list_accounts}
\title{Get the list of accounts from Google Analytics}
\usage{
list_accounts(max.results = NULL, start.index = NULL)
}
\arguments{
\item{max.results}{The maximum number of accounts to include in this response.}

\item{start.index}{An index of the first account to retrieve. Use this parameter as a
pagination mechanism along with the max-results parameter.}
}
\description{
Get the list of accounts from Google Analytics
}

