% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{list_filters}
\alias{list_filters}
\title{Get the list of filters from Google Analytics}
\usage{
list_filters(accountId = "~all", max.results = NULL, start.index = NULL)
}
\arguments{
\item{accountId}{Account ID to retrieve goals for. Can either be a specific account
ID or '~all', which refers to all the accounts that user has access to.}

\item{max.results}{The maximum number of accounts to include in this response.}

\item{start.index}{An index of the first account to retrieve. Use this parameter as a
pagination mechanism along with the max-results parameter.}
}
\description{
Get the list of filters from Google Analytics
}

