% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goals.R
\name{list_goals}
\alias{list_goals}
\title{Get the list of goals from Google Analytics}
\usage{
list_goals(accountId = "~all", webPropertyId = "~all", profileId = "~all",
  max.results = NULL, start.index = NULL)
}
\arguments{
\item{accountId}{Account ID to retrieve goals for. Can either be a specific account
ID or '~all', which refers to all the accounts that user has access to.}

\item{webPropertyId}{Web property ID to retrieve goals for. Can either be a
specific web property ID or '~all', which refers to all the web properties to which the user
has access.}

\item{profileId}{View (Profile) ID to retrieve goals for. Can either be a specific
view (profile) ID or '~all', which refers to all the views (profiles) that user has access to.}

\item{max.results}{The maximum number of accounts to include in this response.}

\item{start.index}{An index of the first account to retrieve. Use this parameter as a
pagination mechanism along with the max-results parameter.}
}
\description{
Get the list of goals from Google Analytics
}

