% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R, R/web_properties.R
\name{list_properties}
\alias{list_properties}
\title{Get the list of profiles from Google Analytics}
\usage{
list_properties(accountId = "~all", max.results = NULL,
  start.index = NULL)

list_properties(accountId = "~all", max.results = NULL,
  start.index = NULL)
}
\arguments{
\item{accountId}{Account ID to retrieve web properties for. Can either be a specific account
ID or '~all', which refers to all the accounts that user has access to.}

\item{max.results}{The maximum number of accounts to include in this response.}

\item{start.index}{An index of the first account to retrieve. Use this parameter as a
pagination mechanism along with the max-results parameter.}

\item{webPropertyId}{Web property ID for the views (profiles) to retrieve. Can either be a
specific web property ID or '~all', which refers to all the web properties to which the user
has access.}

\item{accountId}{Account ID to retrieve web properties for. Can either be a specific account
ID or '~all', which refers to all the accounts that user has access to.}

\item{max.results}{The maximum number of accounts to include in this response.}

\item{start.index}{An index of the first account to retrieve. Use this parameter as a
pagination mechanism along with the max-results parameter.}
}
\description{
Get the list of profiles from Google Analytics

Get the list of properties from Google Analytics
}

