% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_dimensions.R
\name{list_custom_dimensions}
\alias{list_custom_dimensions}
\title{Get the list of custom dimensions from Google Analytics}
\usage{
list_custom_dimensions(accountId, webPropertyId, max.results = NULL,
  start.index = NULL)
}
\arguments{
\item{accountId}{Account ID to retrieve custom metrics for. A specific account
ID needs to be provided.}

\item{webPropertyId}{Web property ID to retrieve custom metrics for. A specific web property ID
needs to be provided.}

\item{max.results}{The maximum number of accounts to include in this response.}

\item{start.index}{An index of the first account to retrieve. Use this parameter as a
pagination mechanism along with the max-results parameter.}
}
\description{
Get the list of custom dimensions from Google Analytics
}

