% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_metrics.R
\name{list_custom_metrics}
\alias{list_custom_metrics}
\title{Get the list of custom metrics from Google Analytics}
\usage{
list_custom_metrics(accountId = "~all", webPropertyId = "~all",
  max.results = NULL, start.index = NULL)
}
\arguments{
\item{accountId}{Account ID to retrieve custom metrics for. Can either be a specific account
ID or '~all', which refers to all the accounts that user has access to.}

\item{webPropertyId}{Web property ID to retrieve custom metrics for. Can either be a
specific web property ID or '~all', which refers to all the web properties to which the user
has access.}

\item{max.results}{The maximum number of results to include in this response.}

\item{start.index}{An index of the first result to retrieve. Use this parameter as a
pagination mechanism along with the max-results parameter.}
}
\description{
Get the list of custom metrics from Google Analytics
}

