% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{list_filters}
\alias{list_filters}
\title{Get the list of filters from Google Analytics}
\usage{
list_filters(accountId, max.results = NULL, start.index = NULL)
}
\arguments{
\item{accountId}{Account ID to retrieve custom metrics for. A specific account
ID needs to be provided.}

\item{max.results}{The maximum number of accounts to include in this response.}

\item{start.index}{An index of the first account to retrieve. Use this parameter as a
pagination mechanism along with the max-results parameter.}
}
\description{
Get the list of filters from Google Analytics
}

